refactor: guest myclass 페이지 리팩토링
