+++ src/pages/guest/page/GuestMyClass/GuestMyClass.style.ts
@@ -62,26 +62,3 @@ export const filterSelectStyle = css`
   margin-top: 1.8rem;
 `;
 
-export const completedTabContainer = (theme: Theme) => css`
-  ${flexGenerator('column')}
-  height: 100dvh;
-  padding: 10.4rem 9.3rem 20.9rem 9.3rem;
-
-  background-color: ${theme.color.background};
-`;
-
-export const textWrapper = css`
-  ${flexGenerator('column', 'center', 'center')}
-  width: 100%;
-  margin-top: 2.2rem;
-`;
-
-export const textStyle = (theme: Theme) => css`
-  color: ${theme.color.lightgray2};
-  ${theme.font['subhead05-sb-14']}
-`;
-
-export const detailWrapper = css`
-  ${flexGenerator('column', 'center', 'center')}
-  gap: 2rem;
-`;
