+++ src/pages/guest/page/GuestMyClass/GuestMyClass.tsx
@@ -1,5 +1,4 @@
-import { Button, FilterSelect, Header, Image } from '@components';
-import GuestMyClassCard from '@pages/guest/components/GuestMyClassCard/GuestMyClassCard';
+import { FilterSelect, Header } from '@components';
 import {
   guestMyClassLayout,
   tapLine,
@@ -8,23 +7,13 @@ import {
   filterSelectWrapper,
   guestMyClassCardContainer,
   filterSelectStyle,
-  completedTabContainer,
-  textWrapper,
-  textStyle,
-  detailWrapper,
 } from './GuestMyClass.style';
 import { useState } from 'react';
 import { GUEST_MY_CLASS_DATA } from 'src/constants/mocks/guestMyClassCardData';
+import { GuestMyClassEmptyView, MoimCard } from '@pages/guest/components';
 
 const GuestMyClass = () => {
   const [activeTab, setActiveTab] = useState<'신청한' | '참가한'>('신청한');
-
-  const appliedClasses = GUEST_MY_CLASS_DATA.filter(
-    (data) => data.moimSubmissionState !== 'COMPLETED'
-  );
-  const completedClasses = GUEST_MY_CLASS_DATA.filter(
-    (data) => data.moimSubmissionState === 'COMPLETED'
-  );
 
   return (
     <>
@@ -40,52 +29,28 @@ const GuestMyClass = () => {
           </div>
         </div>
       </article>
-      {activeTab === '신청한' ? (
-        appliedClasses.length === 0 ? (
-          <article css={completedTabContainer}>
-            <Image width="12rem" src="https://picsum.photos/120" />
-            <div css={detailWrapper}>
-              <div css={textWrapper}>
-                <p css={textStyle}>아직 신청한 모임이 없어요</p>
-                <p css={textStyle}>다양한 클래스 모임을 둘러보세요:{')'}</p>
-              </div>
-              <Button variant="round">클래스 모임 둘러보기</Button>
-            </div>
-          </article>
-        ) : (
-          <div css={guestMyClassCardContainer}>
-            <article css={filterSelectWrapper}>
-              <div css={filterSelectStyle}>
-                <FilterSelect
-                  options={[
-                    '전체',
-                    '입금 대기',
-                    '승인 대기',
-                    '승인 완료',
-                    '승인 거절',
-                    '환불 완료',
-                  ]}
-                />
-              </div>
-            </article>
-            <GuestMyClassCard />
-          </div>
-        )
-      ) : completedClasses.length === 0 ? (
-        <article css={completedTabContainer}>
-          <Image width="12rem" src="https://picsum.photos/120" />
-          <div css={detailWrapper}>
-            <div css={textWrapper}>
-              <p css={textStyle}>아직 참가한 모임이 없어요</p>
-              <p css={textStyle}>다양한 클래스 모임을 둘러보세요:{')'}</p>
-            </div>
-            <Button variant="round">클래스 모임 둘러보기</Button>
+
+      {/* filter select */}
+      {GUEST_MY_CLASS_DATA.length !== 0 && activeTab === '신청한' && (
+        <article css={filterSelectWrapper}>
+          <div css={filterSelectStyle}>
+            <FilterSelect
+              options={['전체', '입금 대기', '승인 대기', '승인 완료', '승인 거절', '환불 완료']}
+            />
           </div>
         </article>
+      )}
+
+      {GUEST_MY_CLASS_DATA.length === 0 ? (
+        <GuestMyClassEmptyView
+          text={activeTab === '신청한' ? '아직 신청한 모임이 없어요' : '아직 참가한 모임이 없어요'}
+        />
       ) : (
-        <div css={guestMyClassCardContainer}>
-          <GuestMyClassCard />
-        </div>
+        GUEST_MY_CLASS_DATA.map((data) => (
+          <div css={guestMyClassCardContainer}>
+            <MoimCard key={data.moimId} guestMyClassData={data} />
+          </div>
+        ))
       )}
     </>
   );
